Versão 1.5 24/11/2021

+++ src/Controller/SQLQuery.sql
@@ -6,14 +6,14 @@ create table tb_user(
 	telefone varchar(100),
 	data_nasc varchar(100),
 	senha varchar (100),
-        saldo float,
+        saldo double,
 	constraint pk_user_cpf primary key (cpf)
 );
 
 create table tb_empresa(
         id serial,
  	nome varchar(255),
- 	valor_atual float,
+ 	valor_atual double,
 	constraint pk_empresa_id primary key (id)
 );
 
@@ -21,7 +21,7 @@ create table tb_ofertas (
 	cpf_user varchar(100),
 	id_empresa integer,
 	quantidade integer,
-	valor float,
+	valor double,
 	constraint fk_user_cpf foreign key (cpf_user) references tb_user(cpf),
         constraint fk_empresa_id foreign key (id_empresa) references tb_empresa(id)
 );
